==> 0_register_dataset.ipynb <==
from azureml.core import Workspace, Dataset, Datastore

### connecting to ML workspace
### --------------------

subscription_id = '8c386bb9-fbb5-45dd-a9cd-2ca847235881'
resource_group = 'rg-aml-ws-ga'
workspace_name = 'aml-ws-ga'

ws = Workspace(subscription_id, resource_group, workspace_name)

### register dataset
### --------------------

datastore = ws.get_default_datastore()
datastore.upload_files(files = ['data/german_credit_dataset.csv'], overwrite = True, show_progress = True)

dataset = Dataset.Tabular.from_delimited_files(path = [(datastore, 'german_credit_dataset.csv')])
dataset.register(ws, name = 'german_credit_hsg', tags = {'purpose': 'demo'}, create_new_version = True)
